refactor: move the masking helper functions to mask.py
